requirements.txt
wheel
pycryptodome>=3.9.8,<3.10
paho-mqtt>=1.5.1,<1.6
pytest>=6.1,<6.2
sphinx>=3.2.1,<3.3
sphinx_rtd_theme
mypy>=0.790
coverage
serial>=0.0.97
pyserial>=3.4
termcolor
PyYAML>=5.3.1,<5.4
python-daemon>=2.2.4, <2.3
